requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
cleo==2.1.0
-e git+ssh://git@github.com/seji-x/BE-Class.git@168ef108d4afd3943ea0441eb20e26145519128f#egg=core
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cryptography==42.0.7
cytoolz==0.12.3
distlib==0.3.9
Django==4.2.7
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dulwich==0.21.7
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
fastjsonschema==2.20.0
filelock==3.16.1
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
importlib_metadata==8.5.0
inflection==0.5.1
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
lru-dict==1.2.0
MarkupSafe==2.1.5
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.0.5
ndg-httpsclient==0.5.1
openapi-codec==1.3.2
packaging==24.0
parsimonious==0.10.0
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.26.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pusher==3.3.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyproject_hooks==1.2.0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rlp==4.0.1
rpds-py==0.18.1
SecretStorage==3.3.3
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tomli==2.0.2
tomlkit==0.13.2
toolz==0.12.1
trove-classifiers==2024.10.16
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.6
web3==6.18.0
websockets==12.0
yarl==1.9.4
zipp==3.20.2
